requirements.txt
brambox==5.0.0
easydict==1.13
matplotlib==3.10.3
numpy==2.3.2
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.3.1
Pillow==11.3.0
scikit_learn==1.7.1
scipy==1.16.0
tensorboardX==2.6.2.2
timm==0.3.2
torchsummary==1.5.1
tqdm==4.65.2
